+++ tests/StringableClass.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace InfluxDB2Test;
+
+class StringableClass
+{
+    public function __toString()
+    {
+        return "stringable";
+    }
+}